refactor: update import paths for purchases and session types
